move device specific to pio device file
