+++ pio_usb.h
@@ -24,3 +24,6 @@ void pio_usb_device_task(void);
 endpoint_t *pio_usb_get_endpoint(usb_device_t *device, uint8_t idx);
 int pio_usb_get_in_data(endpoint_t *ep, uint8_t *buffer, uint8_t len);
 int pio_usb_set_out_data(endpoint_t *ep, const uint8_t *buffer, uint8_t len);
+
+
+void pio_usb_ll_init(pio_port_t *pp, const pio_usb_configuration_t *c, pio_hw_root_port_t* hw_root);
